Updated order model properties with correct data types

+++ Fortnox.NET/Models/Order/Order.cs
@@ -17,17 +17,17 @@ namespace FortnoxNET.Models.Order
         [JsonProperty(PropertyName = "@urlTaxReductionList")]
         public string UrlTaxReductionList { get; set; }
 
-        public float? AdministrationFee { get; set; }
+        public decimal? AdministrationFee { get; set; }
 
         [JsonReadOnly]
-        public float? AdministrationFeeVAT { get; set; }
+        public decimal? AdministrationFeeVAT { get; set; }
 
         public string Address1 { get; set; }
 
         public string Address2 { get; set; }
 
         [JsonReadOnly]
-        public float? BasisTaxReduction { get; set; }
+        public decimal? BasisTaxReduction { get; set; }
 
         [JsonReadOnly]
         public bool? Cancelled { get; set; }
@@ -37,10 +37,10 @@ namespace FortnoxNET.Models.Order
         public string Comments { get; set; }
 
         [JsonReadOnly]
-        public float? ContributionPercent { get; set; }
+        public decimal? ContributionPercent { get; set; }
 
         [JsonReadOnly]
-        public float? ContributionValue { get; set; }
+        public decimal? ContributionValue { get; set; }
 
         public bool? CopyRemarks { get; set; }
 
@@ -48,9 +48,9 @@ namespace FortnoxNET.Models.Order
 
         public string Currency { get; set; }
 
-        public float? CurrencyRate { get; set; }
+        public decimal? CurrencyRate { get; set; }
 
-        public float? CurrencyUnit { get; set; }
+        public decimal? CurrencyUnit { get; set; }
 
         public string CustomerName { get; set; }
 
@@ -78,13 +78,13 @@ namespace FortnoxNET.Models.Order
 
         public string ExternalInvoiceReference2 { get; set; }
 
-        public float? Freight { get; set; }
+        public decimal? Freight { get; set; }
 
         [JsonReadOnly]
-        public float FreightVAT { get; set; }
+        public decimal FreightVAT { get; set; }
 
         [JsonReadOnly]
-        public float? Gross { get; set; }
+        public decimal? Gross { get; set; }
 
         [JsonReadOnly]
         public bool? HouseWork { get; set; }
@@ -93,7 +93,7 @@ namespace FortnoxNET.Models.Order
         public int? InvoiceReference { get; set; }
 
         [JsonReadOnly]
-        public float? Net { get; set; }
+        public decimal? Net { get; set; }
 
         public bool? NotCompleted { get; set; }
 
@@ -125,7 +125,7 @@ namespace FortnoxNET.Models.Order
         public string Remarks { get; set; }
 
         [JsonReadOnly]
-        public float? RoundOff { get; set; }
+        public decimal? RoundOff { get; set; }
 
         [JsonReadOnly]
         public bool? Sent { get; set; }
@@ -138,10 +138,10 @@ namespace FortnoxNET.Models.Order
         public string TermsOfPayment { get; set; }
 
         [JsonReadOnly]
-        public float? Total { get; set; }
+        public decimal? Total { get; set; }
 
         [JsonReadOnly]
-        public float? TotalVAT { get; set; }
+        public decimal? TotalVAT { get; set; }
 
         public bool? VATIncluded { get; set; }
